bound establish_connecion_loop method with a timeout
the total time this method might take is at most the timeout decided in the connection settings, which is the time needed to establish the connection and query for the version

+++ mm2src/coins/utxo/rpc_clients/electrum_rpc/connection_managers/multiple.rs
@@ -140,7 +140,7 @@ impl ConnectionManagerTrait for Arc<ConnectionManagerMultiple> {
                         let Some(client) = manager.get_client() else { break };
                         // This will ping all the active connections, which will keep these connections alive.
                         client.server_ping().compat().await.ok();
-                        Timer::sleep(PING_TIMEOUT_SEC as f64).await;
+                        Timer::sleep(PING_TIMEOUT_SEC).await;
                     }
                 }
             };
